+++ doc/changelog.rst
@@ -80,6 +80,8 @@ Enhancements
 - You can now add ``None`` to a ggplot, doing so returns a copy of the
   the ggplot object. (:issue:`474`)
 
+- Better handling of multiline facet labels. (:issue:`484`)
+
 v0.7.1
 ------
 (2020-08-05)